requirements.txt
fastapi
uvicorn
python-multipart
pillow
requests
tensorflow
tensorflow-hub
sqlalchemy
psycopg2-binary
